requirements.txt.txt
streamlitpandasplotly